refactor: add network map to navbar
