+++ routes/web.php
@@ -8,7 +8,7 @@ use App\Http\Controllers\Inertia\BookmarksController;
 use App\Http\Controllers\Inertia\CompatibleWalletsController;
 use App\Http\Controllers\Inertia\ExchangesController;
 use App\Http\Controllers\Inertia\HomeController;
-use App\Http\Controllers\Inertia\PeersMapController;
+use App\Http\Controllers\Inertia\NetworkMapController;
 use App\Http\Controllers\Inertia\ShowBlockController;
 use App\Http\Controllers\Inertia\ShowTransactionController;
 use App\Http\Controllers\Inertia\StatisticsController;
@@ -97,7 +97,7 @@ Route::post('/exchanges', [ExchangesController::class, 'submit'])
 
 Route::get('/bookmarks', BookmarksController::class)->name('bookmarks');
 
-Route::get('/peers-map', PeersMapController::class)->name('peers-map');
+Route::get('/network-map', NetworkMapController::class)->name('network-map');
 
 Route::post('/webhooks', WebhooksController::class)
     ->withoutMiddleware([VerifyCsrfToken::class])
